update `Char` & refactor `text::char`
- mv `core_impls` to `impls/traits.rs`.
- remove deprecated methods: `len_to_utf8`, `utf8_?bytes_len`.
- return lenghts as `usize`.
- update docs.

+++ libs/base_core/src/text/str/string_nonul.rs
@@ -353,7 +353,7 @@ impl<const CAP: usize> StringNonul<CAP> {
         let mut new = unwrap![ok? Self::new()];
         if c != '\0' {
             let bytes = Char::to_utf8_bytes(c);
-            let len = Char::utf8_len(bytes[0]) as usize;
+            let len = Char::utf8_len(bytes[0]);
             is![CAP < len; return Err(MismatchedCapacity::closed(0, len, CAP))];
             new.arr[0] = bytes[0];
             if len > 1 { new.arr[1] = bytes[1]; }
@@ -393,7 +393,7 @@ impl<const CAP: usize> StringNonul<CAP> {
         let mut new = unwrap![ok? Self::new()];
         if !c.is_nul() {
             let bytes = c.to_utf8_bytes();
-            let len = Char::utf8_len(bytes[0]) as usize;
+            let len = Char::utf8_len(bytes[0]);
             is![CAP < len; return Err(MismatchedCapacity::closed(0, len, CAP))];
             new.arr[0] = bytes[0];
             if len > 1 { new.arr[1] = bytes[1]; }
@@ -414,7 +414,7 @@ impl<const CAP: usize> StringNonul<CAP> {
         let mut new = unwrap![ok? Self::new()];
         if !c.is_nul() {
             let bytes = c.to_utf8_bytes();
-            let len = Char::utf8_len(bytes[0]) as usize;
+            let len = Char::utf8_len(bytes[0]);
             is![CAP < len; return Err(MismatchedCapacity::closed(0, len, CAP))];
             new.arr[0] = bytes[0];
             if len > 1 { new.arr[1] = bytes[1]; }
